+++ RDS.ExpenseTracker.Domain/Models/FinancialAccount.cs
@@ -5,7 +5,7 @@ namespace RDS.ExpenseTracker.Domain.Models
     {
         public int Id { get; set; }
         public string Name { get; set; } = string.Empty;
-        public int Availability { get; set; }
+        public decimal Availability { get; set; }
         public string? Description { get; set; }
     }
 }
